Nuke code that won't be used (for now)

+++ src/nbl/video/utilities/IPropertyPool.cpp
@@ -20,7 +20,7 @@ IPropertyPool::IPropertyPool(uint32_t capacity, void* reserved, bool contiguous)
 {
     if (contiguous)
     {
-        m_indexToAddr = reinterpret_cast<uint32_t*>(reinterpret_cast<uint8_t*>(reserved)+getReservedSize(capacity));
+        m_indexToAddr = reinterpret_cast<uint64_t*>(reinterpret_cast<uint8_t*>(reserved)+getReservedSize(capacity));
         m_addrToIndex = m_indexToAddr+capacity;
 
         std::fill_n(m_indexToAddr,capacity,invalid);
